Add remaining ASL functions. Increase Program Counter in all ASL functions. Remove now-redundant calls of conversion method `toInt()`.

+++ src/main/kotlin/com/cormontia/sim65dd/AddWithCarry.kt
@@ -71,7 +71,7 @@ class AddWithCarry {
         val operand = param.toString(16).uppercase(Locale.getDefault()).padStart(2, '0')
         println("ADC \$$operand, X")
 
-        val location = zeroPageX(cpu, param).toInt()
+        val location = zeroPageX(cpu, param)
         val toAdd = memory[location]
         val result = addWithCary(cpu.acc, toAdd, cpu.C)
         result.applyTo(cpu)
@@ -111,7 +111,7 @@ class AddWithCarry {
         println("ADC \$$msbHex$lsbHex, Y")
 
         val location = absoluteY(cpu, lsb, msb)
-        val toAdd = memory[location.toInt()]
+        val toAdd = memory[location]
         val result = addWithCary(cpu.acc, toAdd, cpu.C)
         result.applyTo(cpu)
 
@@ -123,7 +123,7 @@ class AddWithCarry {
         println("ADC (\$$operand, X)")
 
         val location = indexedIndirectX(cpu, memory, param)
-        val toAdd = memory[location.toInt()]
+        val toAdd = memory[location]
         val result = addWithCary(cpu.acc, toAdd, cpu.C)
         result.applyTo(cpu)
 
@@ -135,7 +135,7 @@ class AddWithCarry {
         println("ADC (\$$operand), Y")
 
         val location = indirectIndexedY(cpu, memory, param)
-        val toAdd = memory[location.toInt()]
+        val toAdd = memory[location]
         val result = addWithCary(cpu.acc, toAdd, cpu.C)
         result.applyTo(cpu)
